feat: histogram page updated & structure of the project purified
