+++ app/components/ProcessUsageBar/ProcessUsageBar.css
@@ -29,4 +29,5 @@
     justify-content: center;
     color: #242c39;
     font-weight: 600;
+    min-width: 12%;
   }
