+++ app/containers/HistoricalAnalysisContainer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import HistoricalAnalysis from '../components/HistoricalAnalysis';
+import HistoricalAnalysis from '../pages/HistoryAnalysis';
 
 export default function HistoricalAnalysisPage() {
   return <HistoricalAnalysis />;
