+++ app/containers/HomeContainer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Home from '../components/Home';
+import Home from '../pages/Home';
 
 export default function HomePage() {
   return <Home />;
