+++ app/pages/HistoryAnalysis/HistoricalAnalysis.css
@@ -1,6 +1,6 @@
 .processCardWrapper {
   overflow-y: scroll;
-  height: calc(100vh - 8.5em);
+  height: calc(100vh - 24em);
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -10,11 +10,10 @@
   position: relative;
   display: flex;
   flex-direction: column;
-  width: 75vw;
-  height: 4em;
+  width: 87.8vw;
+  min-height: 7em;
   margin-top: 0.7em;
   margin-bottom: 0.7em;
-  padding: 2em;
   background-color: #e2e3e5;
   color: #232c39;
   border-radius: 0.2em;
@@ -58,7 +57,7 @@
 .processRow {
   display: flex;
   min-height: 3em;
-  margin-bottom: 1em;
+  margin-top: 1em;
 }
 
 .openHistoricView {
@@ -76,7 +75,6 @@
 
 .processScaleTitleWrapper {
   display: flex;
-  margin-left: -1.8em;
   width: 86.6vw;
   justify-content: space-around;
   position: absolute;
@@ -105,7 +103,9 @@
 .processCardLeftSection {
   display: flex;
   align-items: center;
-  width: 50%;
+  width: 38%;
+  font-weight: 700;
+  margin-left: 2em;
 }
 
 .procesCardRigthSection {
